# Conflicts:
#	DESCRIPTION
#	man/rasterize.Rd

+++ src/rasterize.cpp
@@ -50,7 +50,7 @@ bool LASRrasterize::process(LASpoint*& p)
   else
     cells.push_back(raster.cell_from_xy(x,y));
 
-  for (size_t i = 0 ; i < metrics.size() ; ++i)
+  for (int i = 0 ; i < metrics.size() ; ++i)
   {
     for (int cell : cells)
     {
